Revert "Split pybind file"
This reverts commit c65c8c9915e7dde49e0079ca1e2f5063d99c894c.
