+++ qiskit/providers/aer/backends/wrappers/bindings.cc
@@ -1,18 +1,11 @@
-/**
- * This code is part of Qiskit.
- *
- * (C) Copyright IBM 2018, 2020.
- *
- * This code is licensed under the Apache License, Version 2.0. You may
- * obtain a copy of this license in the LICENSE.txt file in the root directory
- * of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
- *
- * Any modifications or derivative works of this code must retain this
- * copyright notice, and modified files need to carry a notice indicating
- * that they have been altered from the originals.
- */
+#include <iostream>
 
-#include "framework/results/pybind_result.hpp"
+#include <pybind11/pybind11.h>
+
+#include "framework/matrix.hpp"
+#include "framework/types.hpp"
+#include "framework/pybind_json.hpp"
+
 #include "controllers/qasm_controller.hpp"
 #include "controllers/statevector_controller.hpp"
 #include "controllers/unitary_controller.hpp"
